This almost works now.
Broken: the parent_row can't work this simply, probably need an
intermediate data structure to represent this, without it there is not
enough memory for the parent row.

Missing: lines to the words on the right.

Unfortunate: it really is best to grow from the words out, building
"just enough tree", keeping the analysis as close to the words as
possible.